Extract node to a submodule
